+++ src/data/mod.rs
@@ -1,9 +1,15 @@
+mod ast_value;
+mod error;
 mod row;
 mod schema;
 mod table;
 mod value;
 
-pub use row::{Row, RowError};
-pub use schema::Schema;
-pub use table::{get_name, Table, TableError};
-pub use value::{cast_ast_value, is_same_as_data_type_ast_value, Value, ValueError};
+pub use {
+    ast_value::{cast_ast_value, is_same_as_data_type_ast_value},
+    error::DataError,
+    row::{Row, RowError},
+    schema::Schema,
+    table::{get_name, Table, TableError},
+    value::{Value, ValueError},
+};
