+++ src/data/value/mod.rs
@@ -1,4 +1,5 @@
 use {
+    super::DataError,
     crate::result::Result,
     boolinator::Boolinator,
     serde::{Deserialize, Serialize},
@@ -11,8 +12,6 @@ mod error;
 mod group_key;
 mod unique_key;
 
-pub use ast_value::cast_ast_value;
-pub use ast_value::is_same_as_data_type_ast_value;
 pub use error::ValueError;
 
 #[derive(Debug, Clone, Serialize, Deserialize)]
@@ -182,12 +181,12 @@ impl Value {
             (DataType::Boolean, Value::Str(value)) => Ok(match value.to_uppercase().as_str() {
                 "TRUE" => Ok(Value::Bool(true)),
                 "FALSE" => Ok(Value::Bool(false)),
-                _ => Err(ValueError::ImpossibleCast),
+                _ => Err(DataError::ImpossibleCast),
             }?),
             (DataType::Boolean, Value::I64(value)) => Ok(match value {
                 1 => Ok(Value::Bool(true)),
                 0 => Ok(Value::Bool(false)),
-                _ => Err(ValueError::ImpossibleCast),
+                _ => Err(DataError::ImpossibleCast),
             }?),
             (DataType::Boolean, Value::F64(value)) =>
             {
@@ -197,7 +196,7 @@ impl Value {
                 } else if *value == 0.0 {
                     Ok(Value::Bool(false))
                 } else {
-                    Err(ValueError::ImpossibleCast.into())
+                    Err(DataError::ImpossibleCast.into())
                 }
             }
             // Integer
@@ -205,7 +204,7 @@ impl Value {
             (DataType::Int, Value::Str(value)) => Ok(Value::I64(
                 value
                     .parse::<i64>()
-                    .map_err(|_| ValueError::ImpossibleCast)?,
+                    .map_err(|_| DataError::ImpossibleCast)?,
             )),
             (DataType::Int, Value::Bool(value)) => Ok(Value::I64(if *value { 1 } else { 0 })),
             // Float
@@ -216,7 +215,7 @@ impl Value {
             (DataType::Float(_), Value::Str(value)) => Ok(Value::F64(
                 value
                     .parse::<f64>()
-                    .map_err(|_| ValueError::ImpossibleCast)?,
+                    .map_err(|_| DataError::ImpossibleCast)?,
             )),
             (DataType::Text, Value::Bool(value)) => Ok(Value::Str(
                 (if *value { "TRUE" } else { "FALSE" }).to_string(),
@@ -224,7 +223,7 @@ impl Value {
             // Text
             (DataType::Text, Value::I64(value)) => Ok(Value::Str(value.to_string())),
             (DataType::Text, Value::F64(value)) => Ok(Value::Str(value.to_string())),
-            _ => Err(ValueError::UnimplementedCast.into()),
+            _ => Err(DataError::UnimplementedCast.into()),
         }
     }
