+++ src/result.rs
@@ -1,6 +1,6 @@
 use {
     crate::{
-        data::{RowError, TableError, ValueError},
+        data::{DataError, RowError, TableError, ValueError},
         executor::{
             AggregateError, BlendError, CreateTableError, EvaluateError, ExecuteError, FetchError,
             FilterError, JoinError, LimitError, SelectError, UpdateError, ValidateError,
@@ -51,6 +51,8 @@ pub enum Error {
     Validate(#[from] ValidateError),
     #[error(transparent)]
     Value(#[from] ValueError),
+    #[error(transparent)]
+    Data(#[from] DataError),
     #[error(transparent)]
     CreateTable(#[from] CreateTableError),
 }
